Add solutions to exercises and problem sets in Chapter 2.
By now, Chapter 2 is completed.
Learned and coded the algorithms such as:
    - Bubble Sort
    - Insertion Sort
    - Merge Sort
    - Binary Search
Learned to analyse an algorithm in time complexity with notation Θ.
Learned the concept of divide-and-conquer approach, iterative and
recursive method.
Solved 2 problems with Binary Search and Merge Sort algorithms,
sometimes we can use Merge Sort first to sort the array, and that
won't increase the time complexity.